Bugfix: Store Thycotic object in an instance scope variable, not global.
Storing the 'Thycotic' object in a GLOBAL scoped variable is dangerous
in threaded code and causes random failures in Puppet compilations. The
errors manifest as '.has_key?' failures indicating that the '$thycotic'
variable is nil, rather than referring to a Thycotic object like it
should.

Switching to INSTANCE scopign ('@thycotic') makes sure that the
@thycotic variable is filled with a 'Thycotic' object when necessary,
but only for the scope of that application. Furthermore, wrap the init()
procses for that object in a begin/rescue loop so we can get more clear
errors when a failure occurs initializing the object.

Change-Id: I04c39889ee3176f8c1734cc3277bae78b6026821